+++ app/src/main/java/com/spyker3d/tracksnippetplayer/audioplayer/domain/interactor/GetTrackByIdUseCase.kt
@@ -5,5 +5,5 @@ import com.spyker3d.tracksnippetplayer.common.domain.model.Track
 import javax.inject.Inject
 
 class GetTrackByIdUseCase @Inject constructor(private val getTrackByIdRepository: GetTrackByIdRepository) {
-    suspend fun getTrackById(trackId: Int): Track = getTrackByIdRepository.getTrackById(trackId)
+    suspend fun getTrackById(trackId: Long): Track = getTrackByIdRepository.getTrackById(trackId)
 }
