+++ app/src/main/java/com/spyker3d/tracksnippetplayer/audioplayer/domain/repository/GetTrackByIdRepository.kt
@@ -3,5 +3,5 @@ package com.spyker3d.tracksnippetplayer.audioplayer.domain.repository
 import com.spyker3d.tracksnippetplayer.common.domain.model.Track
 
 interface GetTrackByIdRepository {
-    suspend fun getTrackById(trackId: Int): Track
+    suspend fun getTrackById(trackId: Long): Track
 }
